requirements.txt
# =============================================================================
# VectorDB-Bench: Pinned Dependencies for Reproducibility
# =============================================================================
#
# IMPORTANT: All versions are pinned for reproducible benchmarking.
# Last tested: 2024-01 on Python 3.11
#
# Install: pip install -r requirements.txt
# =============================================================================

# -----------------------------------------------------------------------------
# Vector Database Clients (pinned to match Docker images)
# -----------------------------------------------------------------------------
pinecone-client==3.0.2
pymilvus==2.3.4           # Matches milvus:v2.3.4
qdrant-client==1.7.3      # Matches qdrant:v1.7.4
chromadb==0.4.22          # Matches chroma:0.4.22
weaviate-client==4.4.1    # Matches weaviate:1.23.0
pgvector==0.2.4           # Matches pgvector:0.6.0-pg16
psycopg2-binary==2.9.9

# -----------------------------------------------------------------------------
# Core Data & ML
# -----------------------------------------------------------------------------
numpy==1.26.3
pandas==2.1.4
scipy==1.11.4
scikit-learn==1.3.2
torch==2.1.2              # CPU-only for benchmarking
sentence-transformers==2.2.2

# -----------------------------------------------------------------------------
# Datasets
# -----------------------------------------------------------------------------
datasets==2.16.1
huggingface-hub==0.20.2

# Note: BEIR and MTEB have complex dependencies
# We use datasets library directly for BEIR data
# beir>=2.0.0  # Optional, install separately if needed
# mteb>=1.1.0  # Optional, install separately if needed

# -----------------------------------------------------------------------------
# Visualization & Reporting
# -----------------------------------------------------------------------------
matplotlib==3.8.2
seaborn==0.13.1
tabulate==0.9.0

# -----------------------------------------------------------------------------
# Configuration & Utilities
# -----------------------------------------------------------------------------
pyyaml==6.0.1
tqdm==4.66.1
click==8.1.7
marshmallow>=3.0.0,<4.0.0

# -----------------------------------------------------------------------------
# System Monitoring (for cost tracking)
# -----------------------------------------------------------------------------
psutil==5.9.7
docker==7.0.0

# -----------------------------------------------------------------------------
# Development & Testing
# -----------------------------------------------------------------------------
pytest==7.4.4
pytest-cov==4.1.0
black==23.12.1
isort==5.13.2
mypy==1.8.0

# -----------------------------------------------------------------------------
# Jupyter (for analysis notebooks)
# -----------------------------------------------------------------------------
jupyter==1.0.0
ipykernel==6.28.0
